+++ src/tango/sys/consts/fcntl.d
@@ -1,5 +1,0 @@
-deprecated module tango.sys.consts.fcntl;
-
-pragma (msg, "Import tango.stdc.posix.fcntl instead");
-
-public import tango.stdc.posix.fcntl;