+++ RELEASE_NOTES.md
@@ -16,6 +16,21 @@ Removed Symbols
   The (outdated) Makd copy which lived under `script/` has been removed.
 
 
+* `tango.sys.consts.fcntl`
+
+  This module was deprecated since inclusion in ocean (1.25), as it was deprecated
+  by tango 1.3. It is believed to be unused.
+
+* `tango.sys.consts.socket`
+
+  This module was deprecated since inclusion in ocean (1.25), as it was deprecated
+  by tango 1.3. It is believed to be unused.
+
+* `ocean.core.DeepCopy`
+
+  This module was deprecated in ocean 1.24
+
+
 Deprecations
 ============